test: Nickname 클래스 추가에 따른 테스트 코드 변경
