+++ src/test/java/domain/PlayerTest.java
@@ -31,26 +31,6 @@ public class PlayerTest {
         trump = new Trump(strategy);
     }
 
-    @Test
-    @DisplayName("닉네임 처리 테스트: 닉네임이 4~10자인 경우")
-    void 정상_테스트_1() {
-        assertDoesNotThrow(() ->  new Player(new Hand(new ArrayList<>()), Status.HIT, "pobi"));
-    }
-
-    @Test
-    @DisplayName("닉네임 처리 테스트: 닉네임이 4자 미만인 경우")
-    void 예외_테스트_1() {
-        assertThatThrownBy(() ->  new Player(new Hand(new ArrayList<>()), Status.HIT, "pob"))
-            .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    @DisplayName("닉네임 처리 테스트: 닉네임이 10자 초과인 경우")
-    void 예외_테스트_2() {
-        assertThatThrownBy(() ->  new Player(new Hand(new ArrayList<>()), Status.HIT, "jasonjasonj"))
-            .isInstanceOf(IllegalArgumentException.class);
-    }
-
     @Nested
     @DisplayName("승패 판정 테스트")
     class 승패_판정_테스트 {
